# frontend/openapi.json
{
    "openapi": "3.0.0",
    "components": {
...
        "schemas": {
...
            "ErrorResponse": {
                "properties": { "error": { "type": "string" }, "message": { "type": "string" } },
                "required": ["error", "message"],
                "type": "object",
                "additionalProperties": false
            },
...
            "SuccessResponse": {
                "properties": { "message": { "type": "string" } },
                "required": ["message"],
                "type": "object",
                "additionalProperties": false
            },
...
            "LogCreateResponse": {
                "properties": { "id": { "type": "string" } },
                "required": ["id"],
                "type": "object",
                "additionalProperties": false
            },
            "TaskLog": {
                "properties": {
                    "subTaskId": { "type": "string" },
                    "startTime": { "type": "string" },
                    "endTime": { "type": "string" },
                    "done": { "type": "boolean" },
                    "note": { "type": "string" }
                },
                "required": ["subTaskId", "startTime", "endTime", "done"],
                "type": "object",
                "additionalProperties": false
            },
            "LogEntry": {
                "properties": {
                    "id": { "type": "string" },
                    "date": { "type": "string" },
                    "relationshipId": { "type": "string" },
                    "finished": { "type": "boolean" },
                    "closed": { "type": "boolean" },
                    "tasks": { "items": { "$ref": "#/components/schemas/TaskLog" }, "type": "array" }
                },
                "required": ["id", "date", "relationshipId", "finished", "closed", "tasks"],
                "type": "object",
                "additionalProperties": false
            },
            "UpdateLogEntry": {
                "properties": {
                    "date": { "type": "string" },
                    "relationshipId": { "type": "string" },
                    "finished": { "type": "boolean" },
                    "closed": { "type": "boolean" },
                    "tasks": { "items": { "$ref": "#/components/schemas/TaskLog" }, "type": "array" }
                },
                "required": ["date", "relationshipId", "finished", "closed", "tasks"],
                "type": "object",
                "additionalProperties": false
            },
...
        },
        "securitySchemes": { "jwt": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" } }
    },
    "info": { "title": "activity-log-app-backend", "version": "0.0.0", "license": { "name": "ISC" }, "contact": {} },
    "paths": {
...
        "/logs": {
            "post": {
                "operationId": "CreateLog",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        { "$ref": "#/components/schemas/LogCreateResponse" },
                                        { "$ref": "#/components/schemas/ErrorResponse" }
                                    ]
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "content": {
                            "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
                        }
                    }
                },
                "tags": ["Logs"],
                "security": [],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LogEntry" } } }
                }
            },
            "get": {
                "operationId": "GetLogs",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        { "items": { "$ref": "#/components/schemas/LogEntry" }, "type": "array" },
                                        { "$ref": "#/components/schemas/ErrorResponse" }
                                    ]
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "content": {
                            "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
                        }
                    }
                },
                "tags": ["Logs"],
                "security": [],
                "parameters": []
            }
        },
        "/logs/{id}": {
            "get": {
                "operationId": "GetLogById",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        { "$ref": "#/components/schemas/LogEntry" },
                                        { "$ref": "#/components/schemas/ErrorResponse" }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Log not found",
                        "content": {
                            "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "content": {
                            "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
                        }
                    }
                },
                "tags": ["Logs"],
                "security": [],
                "parameters": [{ "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }]
            },
            "put": {
                "operationId": "UpdateLog",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        { "$ref": "#/components/schemas/SuccessResponse" },
                                        { "$ref": "#/components/schemas/ErrorResponse" }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Log not found",
                        "content": {
                            "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "content": {
                            "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
                        }
                    }
                },
                "tags": ["Logs"],
                "security": [],
                "parameters": [{ "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }],
                "requestBody": {
                    "required": true,
                    "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateLogEntry" } } }
                }
            },
            "delete": {
                "operationId": "DeleteLog",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        { "$ref": "#/components/schemas/SuccessResponse" },
                                        { "$ref": "#/components/schemas/ErrorResponse" }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Log not found",
                        "content": {
                            "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "content": {
                            "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
                        }
                    }
                },
                "tags": ["Logs"],
                "security": [],
                "parameters": [{ "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }]
            }
        },
        "/logs/open": {
            "get": {
                "operationId": "GetOpenLogs",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        { "items": { "$ref": "#/components/schemas/LogEntry" }, "type": "array" },
                                        { "$ref": "#/components/schemas/ErrorResponse" }
                                    ]
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "content": {
                            "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
                        }
                    }
                },
                "tags": ["Logs"],
                "security": [],
                "parameters": []
            }
        },
        "/logs/relationship/{recipientId}/{caregiverId}": {
            "get": {
                "operationId": "GetLogsForRelationship",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        { "items": { "$ref": "#/components/schemas/LogEntry" }, "type": "array" },
                                        { "$ref": "#/components/schemas/ErrorResponse" }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Relationship not found",
                        "content": {
                            "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "content": {
                            "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
                        }
                    }
                },
                "tags": ["Logs"],
                "security": [],
                "parameters": [
                    { "in": "path", "name": "recipientId", "required": true, "schema": { "type": "string" } },
                    { "in": "path", "name": "caregiverId", "required": true, "schema": { "type": "string" } }
                ]
            }
        },
...
    },
    "servers": [{ "url": "/" }]
}